% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sarb_annual}
\alias{sarb_annual}
\title{South African Reserve Bank Quarterly Bulletin Data}
\format{
\subsection{\code{sarb_annual}}{

A data frame with approximately 7,240 rows and 60 columns:
\describe{
\item{date}{Date as at start of period}
\item{KBPxxxx}{SARB acronyms for macroeconomic data}
...
}
}
}
\source{
\url{https://www.resbank.co.za}
}
\usage{
sarb_annual
}
\description{
Compilation of annual data
}
\keyword{datasets}
