% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sarb_description}
\alias{sarb_description}
\title{South African Reserve Bank Quarterly Bulletin Data}
\format{
\subsection{\code{sarb_description}}{

A data frame with approximately 7,304 rows and 5 columns:
\describe{
\item{Time_series_number}{SARB acroynm}
\item{Description_and_version}{Description of data}
\item{Version_description}{Version of data}
\item{Frequency}{Frequency of data}
\item{Unit_of_measure}{Units of data}
...
}
}
}
\source{
\url{https://www.resbank.co.za}
}
\usage{
sarb_description
}
\description{
Descriptions of data and acronyms
}
\keyword{datasets}
